tests updated, duplicated global exception handler deleted
