+++ src/test/java/com/dia/ismdtoolbackend/service/CommentServiceImplTest.java
@@ -4,6 +4,7 @@ import com.dia.ismdtoolbackend.entity.CommentEntity;
 import com.dia.ismdtoolbackend.entity.ConceptMetadataEntity;
 import com.dia.ismdtoolbackend.entity.OntologyMetadataEntity;
 import com.dia.ismdtoolbackend.exception.CommentException;
+import com.dia.ismdtoolbackend.exception.CommentNotFoundException;
 import com.dia.ismdtoolbackend.mapper.CommentMapper;
 import com.dia.ismdtoolbackend.models.CommentCreateModel;
 import com.dia.ismdtoolbackend.models.CommentModel;
@@ -368,7 +369,7 @@ class CommentServiceImplTest {
     void deleteComment_CommentNotFound() {
         when(commentRepository.findById(TEST_COMMENT_ID)).thenReturn(Optional.empty());
 
-        CommentException exception = assertThrows(CommentException.class,
+        CommentNotFoundException exception = assertThrows(CommentNotFoundException.class,
                 () -> commentService.deleteComment(TEST_COMMENT_ID));
 
         assertTrue(exception.getMessage().contains("nebyl nalezen"));
